test(document-service): consolidate stale locale relation tests
Move the integration repro next to other document-service relation tests
and drop redundant CM publish scenarios covered by the id-map unit test.
